requirements.txt
sentence-transformers>=5.0.0
faiss-cpu>=1.7.0
pymupdf>=1.23.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
